Refactor: 리뷰사항 반영하여 RenderStars 함수 재사용 가능하도록 분리 및 코드 가독성을 위해 정리
