+++ src/pages/RestaurantDetail/Reviews/Reviews.js
@@ -1,33 +1,13 @@
-import { useCallback } from 'react';
-import FullStar from '../Stars/FullStar';
-import HalfStar from '../Stars/HalfStar';
-import BlankStar from '../Stars/BlankStar';
+import { RenderStars } from '../../../utils/RenderStars';
 import Review from './Review';
 
 export default function Reviews({ reviews, rating }) {
-  const renderStars = useCallback(() => {
-    const stringRating = rating.toString().split('.');
-    let resultArray = [];
-    for (let i = 0; i < Number(stringRating[0]); i++) {
-      resultArray.push(<FullStar key={i} />);
-    }
-    if (stringRating.length > 1)
-      resultArray.push(<HalfStar key={resultArray.length} />);
-    const resultArrayLength = resultArray.length;
-    if (Number(stringRating[0]) < 5) {
-      for (let j = 0; j < 5 - resultArrayLength; j++) {
-        resultArray.push(<BlankStar key={resultArrayLength + j} />);
-      }
-    }
-    return resultArray;
-  }, [rating]);
-
   return (
     <div className="reviews">
       <div className="stars">
         {rating}
         <br />
-        {renderStars()}
+        {RenderStars(rating)}
       </div>
       <div className="review">
         <ul>
